LayoutTests/imported/w3c/web-platform-tests/fetch/range/sw.https.window-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: Error: Invalid Request
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/fetch/range/resources/partial-text.py?action=record-media-range-request&length=100&size=99&partial=206&id=0.5554430010181177&type=audio%2Fmp4.

Harness Error (TIMEOUT), message = null

PASS Defer range header filter tests to service worker
PASS Defer range header passthrough tests to service worker
PASS Ranged response not allowed following no-cors ranged request
PASS Non-opaque ranged response executed
PASS Accept-Encoding should not appear in a service worker
TIMEOUT Opaque range preload successes and failures should be indistinguishable Test timed out
PASS Range headers correctly preserved
PASS Range headers correctly removed
PASS Headers correctly filtered
PASS Include range header in network request
